+++ changes.md
@@ -2,6 +2,15 @@
 
 This document outlines the changes made between versions of the **Goat - SecureMe** firmware.
 
+## V1.3.8
+
+### Changes
+
+#### Pico Network Manager
+
+Updated Pico Network Manager to v1.1.4.
+
+
 ## V1.3.7
 
 ### Changes